# Inkplate 6 Optimizations